requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
Brotli==1.1.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
chromadb==1.1.0
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
ddgs==9.6.0
distro==1.9.0
duckduckgo_search==8.1.1
durationpy==0.10
faiss-cpu==1.12.0
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.9.0
google-ai-generativelanguage==0.7.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==33.1.0
langchain==0.3.27
langchain-chroma==0.2.6
langchain-community==0.3.29
langchain-core==0.3.76
langchain-experimental==0.3.4
langchain-google-genai==2.1.12
langchain-huggingface==0.3.1
langchain-mcp-adapters==0.1.10
langchain-ollama==0.3.8
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.28
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.14.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.3
oauthlib==3.3.1
ollama==0.5.4
onnxruntime==1.22.1
openai==1.107.3
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
pillow==11.3.0
posthog==5.4.0
primp==0.15.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==11.1.1
Pygments==2.19.2
pypdf==6.0.0
PyPika==0.48.9
pyppeteer==2.0.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
typer==0.17.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==1.26.20
uvicorn==0.35.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==10.4
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
